+++ src/misc/wait.ts
@@ -1,0 +1,3 @@
+export async function wait(timeMs: number) {
+    await new Promise<void>(resolve => setTimeout(resolve, timeMs))
+}